ENH: Issue #12 and more work on project settings
- Issue #12 seems to be working well.

- Split mainwin.py up so that each Dock view that makes up the main app
window is in its own python file.

- The project settings dialog code is a mess in the backend and needs to
be refactored (I say that now because I know how I'm going to refactor
it). None of the settings other than the 'trim 0 pressure points' one
works right now, as they all require an active update / action to be
done when the setting changes, which is not supported right now.