Loading hdf5 file reads condition_variables table
If the iohub hdf5 file has a condition_variables table for the session,
it is read into the project, and user can select a start and end time
trial condition variable that will be used to split the file into a set
of T initial segments, where T == the number of trials found. Any pen
data outside of a trial time period is discarded.
If a condition_variables table is not found in the hdf5 file, all pen
samples in file are loaded the same way other supported file formats
are.