ADD: Auto creation of trial segments
- only supported for hdf5 files that have a trial condition variables
table.
- timeline and spatial views limit sample range to the current trial
segment time period.

Known minor issues:

- when a trial segment is selected, all data in the trial is selected
(i.e. timeline's selected data region is set to segment start - end,
which in this case is the full trial.). Not sure what to do about this,
if anything

- when the selected data region is moved in the timeline view, the
'trial segment' becomes deselected in the segment heiarchy. This is
required so that data about the selected region can be displayed in the
properties table. However, this currently also means that you have no
way to see what trial's data is being displayed in the views. This does
need to be addressed, but best way to do so is not clear, so leaving it
as is for now.